[web] Refactor structure in `WalletConnector`
