+++ packages/bento-web/src/components/WalletConnector/connectors/index.ts
@@ -1,0 +1,4 @@
+export { connectKaikas } from './kaikas';
+export { connectKeplr } from './keplr';
+export { connectPhantom } from './phantom';
+export { connectWeb3 } from './web3';
