+++ packages/bento-web/src/components/WalletConnector/index.ts
@@ -1,0 +1,1 @@
+export { WalletConnector } from './WalletConnector';
