+++ packages/bento-web/src/components/WalletConnector/types.ts
@@ -1,0 +1,12 @@
+export type NetworkDraft = {
+  id: string;
+  type: string;
+  name: string;
+  logo: string;
+};
+
+export type ConnectorProps = {
+  networks: NetworkDraft[];
+  signOut: () => Promise<void>;
+  onSave?: () => void;
+};
